refactor: incorporated review comments by refactoring Range class
